+++ app/api/show-api/src/main/java/com/example/artist/controller/dto/response/NumberOfSubscribedArtistApiResponse.java
@@ -1,0 +1,15 @@
+package com.example.artist.controller.dto.response;
+
+
+import com.example.artist.service.dto.response.NumberOfSubscribedArtistServiceResponse;
+
+public record NumberOfSubscribedArtistApiResponse(
+    long count
+) {
+
+    public static NumberOfSubscribedArtistApiResponse from(
+        NumberOfSubscribedArtistServiceResponse response
+    ) {
+        return new NumberOfSubscribedArtistApiResponse(response.count());
+    }
+}
