+++ app/api/show-api/src/main/java/com/example/artist/service/dto/response/NumberOfSubscribedArtistServiceResponse.java
@@ -1,4 +1,4 @@
-package org.example.service.dto.response;
+package com.example.artist.service.dto.response;
 
 public record NumberOfSubscribedArtistServiceResponse(
     long count
