+++ app/api/show-api/src/main/java/com/example/genre/controller/dto/response/NumberOfSubscribedGenreApiResponse.java
@@ -1,0 +1,15 @@
+package com.example.genre.controller.dto.response;
+
+
+import com.example.genre.service.dto.response.NumberOfSubscribedGenreServiceResponse;
+
+public record NumberOfSubscribedGenreApiResponse(
+    long count
+) {
+
+    public static NumberOfSubscribedGenreApiResponse from(
+        NumberOfSubscribedGenreServiceResponse response
+    ) {
+        return new NumberOfSubscribedGenreApiResponse(response.count());
+    }
+}
