+++ app/api/show-api/src/main/java/com/example/genre/service/dto/response/NumberOfSubscribedGenreServiceResponse.java
@@ -1,4 +1,4 @@
-package org.example.service.dto.response;
+package com.example.genre.service.dto.response;
 
 public record NumberOfSubscribedGenreServiceResponse(
     long count
