+++ app/api/show-api/src/main/java/com/example/show/controller/ShowController.java
@@ -1,27 +1,14 @@
 package com.example.show.controller;
 
-import com.example.show.controller.dto.param.ShowAlertPaginationApiParam;
 import com.example.show.controller.dto.param.ShowSearchPaginationApiParam;
-import com.example.show.controller.dto.request.ShowAlertPaginationApiRequest;
-import com.example.show.controller.dto.request.ShowInterestPaginationApiRequest;
 import com.example.show.controller.dto.request.ShowPaginationApiRequest;
 import com.example.show.controller.dto.request.ShowSearchPaginationApiRequest;
-import com.example.show.controller.dto.request.TicketingAlertReservationApiRequest;
-import com.example.show.controller.dto.response.InterestShowPaginationApiResponse;
 import com.example.show.controller.dto.response.ShowDetailApiResponse;
-import com.example.show.controller.dto.response.ShowInterestApiResponse;
 import com.example.show.controller.dto.response.ShowPaginationApiParam;
-import com.example.show.controller.dto.response.TerminatedTicketingShowCountApiResponse;
-import com.example.show.controller.dto.response.TicketingAlertReservationApiResponse;
-import com.example.show.controller.vo.TicketingApiType;
 import com.example.show.service.ShowService;
-import com.example.show.service.dto.param.ShowAlertPaginationServiceParam;
-import com.example.show.service.dto.request.ShowInterestServiceRequest;
 import com.example.show.service.dto.response.ShowPaginationServiceResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.example.dto.response.PaginationApiResponse;
@@ -33,11 +20,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -70,46 +54,6 @@ public class ShowController {
 
     }
 
-    @PostMapping("/{showId}/interests")
-    @Operation(summary = "공연 관심 등록 / 취소")
-    public ResponseEntity<ShowInterestApiResponse> interest(
-        @PathVariable("showId") UUID showId,
-        @AuthenticationPrincipal AuthenticatedUser user
-    ) {
-        return ResponseEntity.ok(
-            ShowInterestApiResponse.from(
-                showService.interest(
-                    ShowInterestServiceRequest.builder()
-                        .showId(showId)
-                        .userId(user.userId())
-                        .build()
-                )
-            )
-        );
-    }
-
-    @GetMapping("/interests")
-    @Operation(summary = "공연 관심 목록 조회")
-    public ResponseEntity<PaginationApiResponse<InterestShowPaginationApiResponse>> getInterests(
-        @ParameterObject ShowInterestPaginationApiRequest request,
-        @AuthenticationPrincipal AuthenticatedUser user
-    ) {
-        var serviceResponse = showService.findInterestShows(
-            request.toServiceRequest(user.userId())
-        );
-
-        List<InterestShowPaginationApiResponse> response = serviceResponse.data().stream()
-            .map(InterestShowPaginationApiResponse::from)
-            .toList();
-
-        return ResponseEntity.ok(
-            PaginationApiResponse.<InterestShowPaginationApiResponse>builder()
-                .data(response)
-                .hasNext(serviceResponse.hasNext())
-                .build()
-        );
-    }
-
     @GetMapping("/{showId}")
     @Operation(summary = "공연 상세 조회")
     public ResponseEntity<ShowDetailApiResponse> getShow(
@@ -121,59 +65,6 @@ public class ShowController {
 
         return ResponseEntity.ok(
             ShowDetailApiResponse.from(showService.getShow(userId, showId, viewIdentifier))
-        );
-    }
-
-    @GetMapping("/{showId}/alert/reservations")
-    @Operation(summary = "공연 티켓팅 알림 예약 조회")
-    public ResponseEntity<TicketingAlertReservationApiResponse> getAlertsReservations(
-        @AuthenticationPrincipal AuthenticatedUser user,
-        @PathVariable("showId") UUID showId,
-        @RequestParam("ticketingApiType") TicketingApiType type
-    ) {
-        return ResponseEntity.ok(
-            TicketingAlertReservationApiResponse.from(
-                showService.findAlertsReservations(user.userId(), showId, type)
-            )
-        );
-    }
-
-    @PostMapping("/{showId}/alert")
-    @Operation(
-        summary = "공연 티켓팅 알림 등록 / 취소",
-        description = "요청한 알람 시간으로 기존 내용을 덮어쓴다."
-    )
-    public ResponseEntity<Void> alert(
-        @AuthenticationPrincipal AuthenticatedUser user,
-        @PathVariable("showId") UUID showId,
-        @RequestParam("ticketingApiType") TicketingApiType type,
-        @Valid @RequestBody TicketingAlertReservationApiRequest ticketingAlertReservationRequest
-    ) {
-        showService.alertReservation(
-            ticketingAlertReservationRequest.toServiceRequest(user.userId(), showId, type)
-        );
-
-        return ResponseEntity.noContent().build();
-    }
-
-    @GetMapping("/alerts")
-    @Operation(summary = "공연 알림 목록 조회")
-    public ResponseEntity<PaginationApiResponse<ShowAlertPaginationApiParam>> getAlerts(
-        @AuthenticationPrincipal AuthenticatedUser user,
-        @ParameterObject ShowAlertPaginationApiRequest request
-    ) {
-        PaginationServiceResponse<ShowAlertPaginationServiceParam> alertShows = showService.findAlertShows(
-            request.toServiceRequest(user.userId()));
-
-        var showAlertPaginationApiParams = alertShows.data().stream()
-            .map(ShowAlertPaginationApiParam::from)
-            .toList();
-
-        return ResponseEntity.ok(
-            PaginationApiResponse.<ShowAlertPaginationApiParam>builder()
-                .data(showAlertPaginationApiParams)
-                .hasNext(alertShows.hasNext())
-                .build()
         );
     }
 
@@ -193,18 +84,6 @@ public class ShowController {
                 .hasNext(response.hasNext())
                 .data(data)
                 .build()
-        );
-    }
-
-    @GetMapping("/terminated/ticketing/count")
-    @Operation(summary = "티켓팅 알림 설정 후 공연이 종료된 개수")
-    public ResponseEntity<TerminatedTicketingShowCountApiResponse> getNumberOfTerminatedTicketingShowCount(
-        @AuthenticationPrincipal AuthenticatedUser user
-    ) {
-        return ResponseEntity.ok(
-            TerminatedTicketingShowCountApiResponse.from(
-                showService.countTerminatedTicketingShow(user.userId())
-            )
         );
     }
 }
