+++ app/api/show-api/src/main/java/com/example/show/controller/dto/usershow/response/NumberOfInterestShowApiResponse.java
@@ -1,0 +1,15 @@
+package com.example.show.controller.dto.usershow.response;
+
+
+import com.example.show.service.dto.usershow.response.NumberOfInterestShowServiceResponse;
+
+public record NumberOfInterestShowApiResponse(
+    long count
+) {
+
+    public static NumberOfInterestShowApiResponse from(
+        NumberOfInterestShowServiceResponse response
+    ) {
+        return new NumberOfInterestShowApiResponse(response.count());
+    }
+}
