+++ app/api/show-api/src/main/java/com/example/show/controller/dto/usershow/response/NumberOfTicketingAlertApiResponse.java
@@ -1,0 +1,15 @@
+package com.example.show.controller.dto.usershow.response;
+
+
+import com.example.show.service.dto.usershow.response.NumberOfTicketingAlertServiceResponse;
+
+public record NumberOfTicketingAlertApiResponse(
+    long count
+) {
+
+    public static NumberOfTicketingAlertApiResponse from(
+        NumberOfTicketingAlertServiceResponse response
+    ) {
+        return new NumberOfTicketingAlertApiResponse(response.count());
+    }
+}
