+++ app/domain/user-domain/src/main/java/org/example/usecase/admin/AdminUseCase.java
@@ -1,4 +1,4 @@
-package org.example.usecase;
+package org.example.usecase.admin;
 
 import lombok.RequiredArgsConstructor;
 import org.example.entity.Admin;
