+++ app/domain/user-domain/src/main/java/org/example/usecase/user/UserUseCase.java
@@ -1,4 +1,4 @@
-package org.example.usecase;
+package org.example.usecase.user;
 
 import java.util.NoSuchElementException;
 import java.util.UUID;
